adding suggestions, renaming boxes
