+++ boxes/README.md
@@ -1,5 +1,28 @@
-# Aztec Starter Kits
+# Aztec Boxes
 
-A collection of projects that act as starter kits for Aztec.
-These depend on accounts, aztec.js and aztec.nr, and point to the monorepos local development versions.
-These projects depend on the sandbox. They're built and tested after the sandbox by docker compose files.
+Aztec Boxes are the one-stop-shop for developing on Aztec. They often include a combination of:
+
+- Fully tested contracts
+- Frontend boilerplates
+- General tests
+
+Boxes include the sandbox installation script and its start command. By choosing the appropriate box, you can get started working on Aztec in a minimal amount of time.
+
+## Getting started
+
+If you have [node](https://nodejs.org/en/download) installed, you can open a terminal in any folder and run:
+
+`npx create-aztec-app`
+
+The script will install the sandbox, run it, and clone the boilerplate you chose. If at any time you encounter problems, refer to the guides at [docs.aztec.network](https://docs.aztec.network) for more information.
+
+## Templates
+
+Currently there are two boxes:
+
+- React - A React boilerplate with a minimal UI.
+- Vanilla JS and HTML - Some say if you get something working in vanilla JS and HTML, you can make it work on any framework. If you can't find the box you need, this could be a good starting point.
+
+## Support
+
+Need any support? Reach out on [discord](https://discord.gg/DgWG2DBMyB), [discourse](https://discourse.aztec.network/) or [twitter](https://twitter.com/aztecnetwork).
