--- boxes/blank-react/README.md
@@ -1,31 +1,0 @@
-# Aztec Box
-
-This box is a one-stop-shop for Aztec that will deploy a blank React Page. You can use it as a boilerplate to start developing your own Aztec app in seconds!
-
-## Prerequisites
-
-- You should have Docker installed. If you don't, follow [this guide](https://docs.aztec.network/dev_docs/getting_started/quickstart#install-docker).
-
-## Installation
-
-Run `npx create-aztec-app` and follow the instructions to clone this repo, install the sandbox and run it on port 8080.
-
-### 1. Install dependencies
-
-Run:
-
-```bash
-yarn
-```
-
-Time to build. Run:
-
-```bash
-yarn dev
-```
-
-Your React app is waiting for you on port 5176. Time to make it your own!
-
-[Read the full Sandbox reference](https://docs.aztec.network/dev_docs/cli/sandbox-reference) for more info on what exactly is happening on your machine!
-
-Visit the [Aztec Docs](https://docs.aztec.network) for more information on how Aztec works, and the [Awesome Aztec Repository](https://github.com/AztecProtocol/awesome-aztec) for more cool projects, boilerplates and tooling.
