+++ boxes/docker-compose.yml
@@ -25,4 +25,4 @@ services:
       ETHEREUM_HOST: http://ethereum:8545
       CHAIN_ID: 31337
       PXE_URL: http://aztec:8080
-      BOX: ${BOX:-box-blank}
+      BOX: ${BOX:-box-vanilla}
