+++ boxes/react/.eslintrc.cjs
