+++ boxes/react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@aztec/box-blank-react",
+  "name": "@aztec/box-react",
   "private": true,
   "version": "0.1.0",
   "type": "module",
