+++ boxes/react/src/hooks/useContract.tsx
@@ -1,12 +1,12 @@
-import { useState } from "react";
-import { deployerEnv } from "../config";
+import { useState } from 'react';
+import { deployerEnv } from '../config';
 
-import { Contract, ContractDeployer, Fr } from "@aztec/aztec.js";
-import { BlankContract } from "../../artifacts/Blank";
-import { toast } from "react-toastify";
+import { Contract, ContractDeployer, Fr } from '@aztec/aztec.js';
+import { BoxReactContract } from '../../artifacts/BoxReact';
+import { toast } from 'react-toastify';
 
 export function useContract() {
-  const { artifact, at } = BlankContract;
+  const { artifact, at } = BoxReactContract;
   const [wait, setWait] = useState(false);
   const [contract, setContract] = useState<Contract | undefined>();
 
@@ -22,11 +22,11 @@ export function useContract() {
       .deploy(Fr.random(), wallet.getCompleteAddress().address)
       .send({ contractAddressSalt: salt });
     const { contractAddress } = await toast.promise(tx.wait(), {
-      pending: "Deploying contract...",
+      pending: 'Deploying contract...',
       success: {
         render: ({ data }) => `Address: ${data.contractAddress}`,
       },
-      error: "Error deploying contract",
+      error: 'Error deploying contract',
     });
 
     const deployerWallet = await deployerEnv.getWallet();
