+++ boxes/react/src/hooks/useNumber.tsx
