+++ boxes/react/src/pages/contract.tsx
