+++ boxes/react/src/pages/home.tsx
