+++ boxes/react/webpack.config.js
