+++ boxes/vanilla-js/.eslintrc.cjs
