+++ boxes/vanilla-js/.prettierignore
