+++ boxes/vanilla-js/.prettierrc.json
