+++ boxes/vanilla-js/src/config.ts
@@ -1,5 +1,5 @@
 import { createPXEClient, waitForPXE } from '@aztec/aztec.js';
-import { BlankContractArtifact } from '../artifacts/Blank.js';
+import { VanillaContractArtifact } from '../artifacts/Vanilla';
 
 class PXE {
   pxeUrl;
@@ -25,5 +25,5 @@ class PXE {
 }
 
 export const pxe = new PXE();
-export const contractArtifact = BlankContractArtifact;
+export const contractArtifact = VanillaContractArtifact;
 export const CONTRACT_ADDRESS_PARAM_NAMES = ['address'];
