+++ boxes/vanilla-js/src/contracts/src/main.nr
@@ -1,4 +1,4 @@
-contract Blank {
+contract Vanilla {
     use dep::aztec::{
         protocol_types::address::AztecAddress,
         state_vars::{singleton::Singleton, map::Map},
