+++ boxes/vanilla-js/tsconfig.json
