+++ boxes/vanilla-js/webpack.config.js
